requirements.txt
google-generativeai
speechrecognition
gtts
playsound
pillow
